server/requirements.txt
fastapi
langchain
youtube_transcript_api
pymongo
python-dotenv
langchain_groq
pillow
langchain_community
nltk
PyPDF2
validators
pyjwt
python-multipart
beautifulsoup4  
tldextract
werkzeug
uvicorn[standard] 
phidata
duckduckgo-search
google-generativeai
google.genai
PyMuPDF
httpx 
sentry-sdk
langchain-mcp-adapters
redis
fastapi-limiter
stripe
resend
livekit
livekit-agents
livekit-plugins-silero
livekit-plugins-deepgram
markitdown[all]